Add option to allow debugging Android services as well
Summary:
In React Native for VR we use `FlipperUtils.shouldEnableFlipper()` to determine whether Flipper should be enabled for the given app build.

The problem is that some of our apps are legitimately running as services (e.g. `com.oculus.explore:explore`), so the check added inside Flipper in D5126205 which only enables Flipper for "main application's process".

This diff adds an ability to relax this requirement (as we do want to do it for the RN VR applications)

Differential Revision: D44511667

fbshipit-source-id: b190d1b70e1375cd957a2c8c542c421701bf12ed