Sync login atom to prescense of token
Summary:
Logged in state now means that we have a valid token. On startup we try to retrieve a token, if its there then we are logged in, if we remove the tokem, due to explicit logout, or if its expired we are logged out.

Connected is more dynamic, it means we can actually hit intern with a sucessful response. It implicitly requires a token / being logged in

Reviewed By: aigoncharov

Differential Revision: D44502482

fbshipit-source-id: e3077101766cba5128a61d62be3bbd1ca1f00b4f