Remove unused plugin update command / redux action
Summary:
This command is never actually dispatched so removing to avoid future confusion.

When new plugins are downloaded a check is done here https://fburl.com/code/e3zedljb to see if pluginIsDisabledForAllConnectedClients, if so, load command is queued, which internally calls update (which resets state)

If the plugin is not disabled then a message appears asking user to manually reload.

Reviewed By: aigoncharov

Differential Revision: D44502476

fbshipit-source-id: b796b77f843218c07ee612d9f45cacba073362d3