Ignore OverlayHandlerView as active child
Summary:
On foldable devices, there's an issue whereas the topmost root view is of type `OverlayHandlerView`.

This doesn't seem to be related in any way to the running app's UI. Or at least is not something of interest when debugging the running app's UI (for example, ig4a).

This change effectively ignores the `OverlayHandlerView` instance as active child and instead returns the next to last root view.

Reviewed By: LukeDefeo

Differential Revision: D44579223

fbshipit-source-id: cea52289c5ba1e6a96817232ac54725b8d2f48d9