limit vscode code action apply to eslint only
Summary:
I accidentally used vscode. I acctidentally used an await keyword in a sync funciton. typescript has an codeaction to add an async to the function which was out of my screen. This changed the function return type to promise which caused other errors around the codebase and it took me a while to figure out where the error was coming from.

All eslint code actions are safe to apply which is not the case with typescript. Thus making this change.

Reviewed By: lblasa

Differential Revision: D45773531

fbshipit-source-id: 1f8835c442d7703644b1af729ab55fcef7d5be73