Make sure FlipperSocketImpl always sets thread TrafficStats tag
Summary:
When looking at some of our logcats in logview, I noticed lots of StrictMode warnings related to a WebSocket not being tagged, and I figured out it was due to FlipperSocketImpl.

By making sure TrafficStats stats tag is set no matter if the SSL factory is used or not, it reduces the amount of StrictMode warnings we see.

Related diff: D38280819

Differential Revision: D46024367

fbshipit-source-id: 92a0bfcef8698a61ef9b27949a5bc0859e03448f