Remove broken img from the main page
Summary:
There's a picture with a broken link in the hero section:
{F999641292}

I checked a list of existing static images and found one which was on this place before: https://fburl.com/code/yh73399d. But the image no longer align with others, since it has low resolution and an old version of the Flipper interface, so I decided to simply remove the broken image from the page.

Reviewed By: passy

Differential Revision: D46074714

fbshipit-source-id: 186b17b9b992f123fc667c701b4f49d9e3cbbb2d