Bootstrap logging and performance improvement
Summary:
Add a few console logs as to highlight how long does it take for our server to start.

Additionally, do not wait until we are ready for client connections before attempting to launch. There's nothing else we do at this point. Instead, launch and wait until we are ready.

Reviewed By: passy

Differential Revision: D46639482

fbshipit-source-id: c098d229edc3cd52e5c876c509a7c81532635afa