Add link to Flipper GraphQL plugin
Summary:
This just adds a link next to each request in the Flipper GraphQL plugin to take you to the mobile GraphQL efficiency tool. We want this to be a kind of hub for mobile devs using GraphQL, alongside GraphiQL.

This was a suggestion from a product dev who attended our demo. Later we could also pull down actually efficiency info and show it directly in Flipper.

Reviewed By: ginfung

Differential Revision: D47029548

fbshipit-source-id: 05185e57bfbfaab9169a6b154822ede82d50fd4d