Bump SocketRocket 0.7.0
Summary:
About to release React Native 0.72.1. This version of React Native has to depend on SocketRocket 0.7.0, released yesterday.

Cocoapods is failing to install dependencies with the following error:
```
In snapshot (Podfile.lock):
    SocketRocket (= 0.7.0)

  In Podfile:
    FlipperKit/FlipperKitReactPlugin (= 0.182.0) was resolved to 0.182.0, which depends on
      FlipperKit/Core (= 0.182.0) was resolved to 0.182.0, which depends on
        SocketRocket (~> 0.6.0)

    React-Core/RCTWebSocket (from `../react-native/`) was resolved to 1000.0.0, which depends on
      SocketRocket (= 0.7.0)

    React-CoreModules (from `../react-native/React/CoreModules`) was resolved to 1000.0.0, which depends on
      SocketRocket
```

By bumping the SocketRocket dependencies and publishing a new version of FlipperKit, we would be able to release React Native.

Reviewed By: passy, cipolleschi

Differential Revision: D47125059

fbshipit-source-id: d0797880c502c14db9f0261c3e8abf2652e7bba2