Enable voice over mode on device while running in accessibility exploration mode
Summary:
All the FoA have different logic depending on whether device is running in  VoiceOver mode or not. This results in incorrect accessibility being reported through the Flipper plugin.

For example on Instagram we fetch the image descriptions, in Facebook app we group individual elements into the accessibility groups to simplify traversing the screen

This tells device that it is in a voice over mode while inspecting accessibility mode is on meaning that it will be the closest approximation to the

Reviewed By: nscoding

Differential Revision: D49579241

fbshipit-source-id: f3be9057007ee6eefbe3b448f0f97c1fcd2c8f46